SubChannelSender transmission across SubChannel
Use tracing crate to ease debugging.

Error handling
* Prefer errors to panics
* Always log an error before returning an InternalError

Note that bincode serialises a SubChannelSender
twice in some situations in order to determine
the serialised size.

Test multiplexing not running out of file descriptors

Implement SubChannelId in terms of Uuid. This avoids
communication to allocate a new SubChannelId, which
simplifies the API.